pyproject.toml
[tool.poetry]
name = "weni-cli"
version = "2.0.0a9"
description = ""
authors = ["Paulo Bernardo <paulo.bernardo@weni.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.8"
requests = "^2.32.3"
flask = "^3.1.0"
waitress = "^3.0.2"
pyyaml = "^6.0.2"
python-slugify = "^8.0.4"
regex = "^2024.11.6"
weni-agents-toolkit = "0.3.0a0"
rich-click = "^1.8.6"

[tool.poetry.dev-dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"
pytest-cov = "^6.0.0"
setuptools = "^75.7.0"
twine = "^6.0.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.3"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}

[tool.poetry.scripts]
weni = "weni_cli.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.mypy]
disable_error_code = ["import-untyped"]
